predictive-maintenance/maintenance-data/requirements.txt
psycopg2-binary>=2.9.0
pathlib2>=2.3.0
